Move the reduxHooks file to the features folder.
