+++ src/Components/CustomHooks/ResetCharacterHook.tsx
@@ -1,4 +1,4 @@
-import { useAppDispatch } from 'Components/CustomHooks/reduxHooks';
+import { useAppDispatch } from 'features/reduxHooks';
 import { resetSelectedOptions } from 'features/selectedOptionsSlice';
 import { resetCharacter } from 'features/characterSlice';
