+++ src/Components/Display Attribute Cards/DisplayAdvantages.tsx
@@ -1,5 +1,5 @@
 import AdvantagesArray from 'Attribute Objects/Advantages';
-import { useAppSelector } from 'Components/CustomHooks/reduxHooks';
+import { useAppSelector } from 'features/reduxHooks';
 import React, { useEffect, useState } from 'react';
 import { AttributeObj } from '../../../types';
 import { DisplayAttributeCards } from './DisplayAttributeCards';
